Simple idling implementation
The main loops are constantly running and using a lot of CPU for
nothing. This implements a simple idling algorithm:

- If there are no user inputs, and
- Nothing already asked to wait for retrace,

Then it should be okay to sleep for 10ms (not more, because DIV timers
are based on a 10ms system clock).

This may not be perfect: adding "something_happened = 1" in more places
may be needed. But at least DIV runs without my laptop fans spinning
like crazy now.